[Edit]:MainView 작업 SplashView로 이동 작업 #128
